chore(origin-ui-core): combine claiming and bulk claiming into one modal
